@@ -19,7 +19,7 @@ documentation = "https://buildingmotif.readthedocs.io/"
 buildingmotif = 'buildingmotif.bin.cli:app'
 
 [tool.poetry.dependencies]
-python = ">=3.9, <3.12"
+python = ">=3.10, <3.13"
 rdflib = ">=7.0"
 SQLAlchemy = {extras = ["mypy"], version = "^1.4.44"}
 pyaml = "^21.10.1"
@@ -34,7 +34,7 @@ Flask-API = "^3.0.post1"
 rfc3987 = "^1.3.8"
 setuptools = "^73.0.1"
 psycopg2 = {version="^2.9.5", optional=true}
-pygit2 = "~1.11.1"
+pygit2 = "~1.17"
 jsonschema = "^4.21.1"
 brick-tq-shacl = {optional = true, version="0.3.3a3"}
 werkzeug="^2.3.7"
@@ -42,6 +42,7 @@ types-jsonschema = "^4.21.0.20240311"
 pandas = "^2.2.2"
 scipy = "<1.14"
 numpy = "2.0"
+flake8 = "^7.1.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.3.0"
@@ -56,7 +57,7 @@ jupyter = "^1.0.0"
 ipykernel = "^6.13.0"
 rise = "^5.7.1"
 jupyter-book = "^0.15.1"
-flake8 = "^5.0.0"
+flake8 = "^7.0.0"
 BAC0 = "^22.9.21"
 netifaces = "^0.11.0"
 pytz = "^2022.7.1"
