@@ -2,22 +2,17 @@ import logging
 from typing import List
 
 import numpy as np
-from rdflib import Namespace
+from rdflib import Graph
 
 from buildingmotif.dataclasses import Library
-from buildingmotif.semantic_graph_synthesizer.bipartite_token_mapper import (
-    BipartiteTokenMapper,
-)
-from buildingmotif.semantic_graph_synthesizer.classes import (
+from buildingmotif.graph_generation.bipartite_token_mapper import BipartiteTokenMapper
+from buildingmotif.graph_generation.classes import (
     Bindings,
     Cost,
     LabelSet,
     Token,
     TokenizedLabel,
 )
-
-BLDG = Namespace("urn:building/")
-
 
 logger = logging.getLogger()
 
@@ -68,7 +63,9 @@ class SemanticGraphSynthesizer:
 
         return labelsets
 
-    def find_bindings_for_label(self, label: TokenizedLabel) -> Bindings:
+    def find_bindings_for_label(
+        self, ontology: Graph, label: TokenizedLabel
+    ) -> Bindings:
         """Gets the bindings for a specific label."""
         best_bindings = Bindings(
             label=label, template=None, bindings={}, cost=Cost.inf()
@@ -83,7 +80,7 @@ class SemanticGraphSynthesizer:
             # compute the best bindings for using the tokens of the label with the template
             # and the cost of the bindings
             bindings, cost = BipartiteTokenMapper.find_bindings_for_tokens_and_template(
-                label.tokens, template
+                ontology, label.tokens, template
             )
             logger.debug(f"- {template.name} {cost.scalar}")
 
@@ -103,7 +100,9 @@ class SemanticGraphSynthesizer:
 
         return best_bindings
 
-    def find_bindings_for_labelset(self, labelset: LabelSet) -> List[Bindings]:
+    def find_bindings_for_labelset(
+        self, ontology: Graph, labelset: LabelSet
+    ) -> List[Bindings]:
         """Find the bindings a given LabelSet."""
         index_label = TokenizedLabel(
             label="Index Label",
@@ -112,7 +111,10 @@ class SemanticGraphSynthesizer:
                 for i, tc in enumerate(labelset.token_classes)
             ],
         )
-        index_bindings = self.find_bindings_for_label(index_label)
+        index_bindings = self.find_bindings_for_label(ontology, index_label)
+        logger.debug(
+            f"Index bindings: {index_bindings.bindings}. Now going through labels"
+        )
 
         bindings_list = []
         for label in labelset.labels:
@@ -124,6 +126,7 @@ class SemanticGraphSynthesizer:
                 )
                 for param, token in index_bindings.bindings.items()
             }
+            logger.debug(f"Bindings for label {label.label}: {bindings}")
 
             bindings_list.append(
                 Bindings(
@@ -133,10 +136,13 @@ class SemanticGraphSynthesizer:
                     cost=index_bindings.cost,
                 )
             )
+            logger.debug(f"Added binding: {bindings_list[-1]}")
 
         return bindings_list
 
-    def find_bindings_for_labels(self, labels: List[TokenizedLabel]) -> List[Bindings]:
+    def find_bindings_for_labels(
+        self, ontology: Graph, labels: List[TokenizedLabel]
+    ) -> List[Bindings]:
         """Find the bindings a given labels.
 
         Groups them in label sets for optimization purposes.
@@ -145,6 +151,9 @@ class SemanticGraphSynthesizer:
 
         bindings_list = []
         for labelset in labelsets:
-            bindings_list.extend(self.find_bindings_for_labelset(labelset))
+            logger.debug(
+                f"Generating bindings for labelset: {labelset.token_classes} {labelset.labels}"
+            )
+            bindings_list.extend(self.find_bindings_for_labelset(ontology, labelset))
 
         return bindings_list
